# src/Api/swagger.json
{
      "openapi": "3.0.4",
      "info": {
            "title": "Metro do Porto API",
            "version": "v1"
      },
      "paths": {
...
            "/v1/porto/metro/route-with-trips/{routeId}": {
                  "get": {
                        "tags": [
                              "Metro"
                        ],
                        "parameters": [
                              {
                                    "name": "routeId",
                                    "in": "path",
                                    "required": true,
                                    "schema": {
                                          "type": "string"
                                    }
                              }
                        ],
                        "responses": {
                              "200": {
                                    "description": "OK",
                                    "content": {
                                          "text/plain": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/RouteWithTripsDto"
                                                }
                                          },
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/RouteWithTripsDto"
                                                }
                                          },
                                          "text/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/RouteWithTripsDto"
                                                }
                                          }
                                    }
                              }
                        }
                  }
            },
            "/v1/porto/metro/trip-with-stops/{tripId}": {
                  "get": {
                        "tags": [
                              "Metro"
                        ],
                        "parameters": [
                              {
                                    "name": "tripId",
                                    "in": "path",
                                    "required": true,
                                    "schema": {
                                          "type": "string"
                                    }
                              }
                        ],
                        "responses": {
                              "200": {
                                    "description": "OK",
                                    "content": {
                                          "text/plain": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/TripWithStopTimesDto"
                                                }
                                          },
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/TripWithStopTimesDto"
                                                }
                                          },
                                          "text/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/TripWithStopTimesDto"
                                                }
                                          }
                                    }
                              }
                        }
                  }
            },
            "/v1/porto/metro/upcoming-departures/{stopId}": {
                  "get": {
                        "tags": [
                              "Metro"
                        ],
                        "parameters": [
                              {
                                    "name": "stopId",
                                    "in": "path",
                                    "required": true,
                                    "schema": {
                                          "type": "string"
                                    }
                              },
                              {
                                    "name": "referenceTime",
                                    "in": "query",
                                    "schema": {
                                          "type": "string",
                                          "format": "date-time"
                                    }
                              }
                        ],
                        "responses": {
                              "200": {
                                    "description": "OK",
                                    "content": {
                                          "text/plain": {
                                                "schema": {
                                                      "type": "array",
                                                      "items": {
                                                            "$ref": "#/components/schemas/StopTimeWithStopDto"
                                                      }
                                                }
                                          },
                                          "application/json": {
                                                "schema": {
                                                      "type": "array",
                                                      "items": {
                                                            "$ref": "#/components/schemas/StopTimeWithStopDto"
                                                      }
                                                }
                                          },
                                          "text/json": {
                                                "schema": {
                                                      "type": "array",
                                                      "items": {
                                                            "$ref": "#/components/schemas/StopTimeWithStopDto"
                                                      }
                                                }
                                          }
                                    }
                              }
                        }
                  }
            },
...
            "/v1/porto/metro/trips/{id}": {
                  "get": {
                        "tags": [
                              "Trips"
                        ],
                        "parameters": [
                              {
                                    "name": "id",
                                    "in": "path",
                                    "required": true,
                                    "schema": {
                                          "type": "string"
                                    }
                              }
                        ],
                        "responses": {
                              "200": {
                                    "description": "OK",
                                    "content": {
                                          "text/plain": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/Trip"
                                                }
                                          },
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/Trip"
                                                }
                                          },
                                          "text/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/Trip"
                                                }
                                          }
                                    }
                              }
                        }
                  }
            },
            "/v1/porto/metro/trips/route/{routeId}": {
                  "get": {
                        "tags": [
                              "Trips"
                        ],
                        "parameters": [
                              {
                                    "name": "routeId",
                                    "in": "path",
                                    "required": true,
                                    "schema": {
                                          "type": "string"
                                    }
                              }
                        ],
                        "responses": {
                              "200": {
                                    "description": "OK",
                                    "content": {
                                          "text/plain": {
                                                "schema": {
                                                      "type": "array",
                                                      "items": {
                                                            "$ref": "#/components/schemas/Trip"
                                                      }
                                                }
                                          },
                                          "application/json": {
                                                "schema": {
                                                      "type": "array",
                                                      "items": {
                                                            "$ref": "#/components/schemas/Trip"
                                                      }
                                                }
                                          },
                                          "text/json": {
                                                "schema": {
                                                      "type": "array",
                                                      "items": {
                                                            "$ref": "#/components/schemas/Trip"
                                                      }
                                                }
                                          }
                                    }
                              }
                        }
                  }
            }
      },
      "components": {
            "schemas": {
...
                  "Route": {
                        "type": "object",
                        "properties": {
                              "id": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "routeId": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "agencyId": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "routeShortName": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "routeLongName": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "routeDesc": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "routeType": {
                                    "type": "integer",
                                    "format": "int32"
                              },
                              "routeUrl": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "routeColor": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "routeTextColor": {
                                    "type": "string",
                                    "nullable": true
                              }
                        },
                        "additionalProperties": false
                  },
                  "RouteWithTripsDto": {
                        "type": "object",
                        "properties": {
                              "route": {
                                    "$ref": "#/components/schemas/Route"
                              },
                              "trips": {
                                    "type": "array",
                                    "items": {
                                          "$ref": "#/components/schemas/Trip"
                                    },
                                    "nullable": true
                              }
                        },
                        "additionalProperties": false
                  },
...
                  "Stop": {
                        "type": "object",
                        "properties": {
                              "id": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "stopId": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "stopCode": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "stopName": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "stopDesc": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "stopLat": {
                                    "type": "number",
                                    "format": "double"
                              },
                              "stopLon": {
                                    "type": "number",
                                    "format": "double"
                              },
                              "zoneId": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "stopUrl": {
                                    "type": "string",
                                    "nullable": true
                              }
                        },
                        "additionalProperties": false
                  },
                  "StopTime": {
                        "type": "object",
                        "properties": {
                              "id": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "tripId": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "arrivalTime": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "departureTime": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "stopId": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "stopSequence": {
                                    "type": "integer",
                                    "format": "int32"
                              },
                              "stopHeadsign": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "pickupType": {
                                    "type": "integer",
                                    "format": "int32",
                                    "nullable": true
                              },
                              "dropOffType": {
                                    "type": "integer",
                                    "format": "int32",
                                    "nullable": true
                              },
                              "shapeDistTraveled": {
                                    "type": "number",
                                    "format": "double",
                                    "nullable": true
                              }
                        },
                        "additionalProperties": false
                  },
                  "StopTimeWithStopDto": {
                        "type": "object",
                        "properties": {
                              "stopTime": {
                                    "$ref": "#/components/schemas/StopTime"
                              },
                              "stop": {
                                    "$ref": "#/components/schemas/Stop"
                              }
                        },
                        "additionalProperties": false
                  },
...
                  "Trip": {
                        "type": "object",
                        "properties": {
                              "id": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "routeId": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "serviceId": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "tripId": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "tripHeadsign": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "directionId": {
                                    "type": "integer",
                                    "format": "int32",
                                    "nullable": true
                              },
                              "blockId": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "shapeId": {
                                    "type": "string",
                                    "nullable": true
                              }
                        },
                        "additionalProperties": false
                  },
                  "TripWithStopTimesDto": {
                        "type": "object",
                        "properties": {
                              "trip": {
                                    "$ref": "#/components/schemas/Trip"
                              },
                              "stopTimes": {
                                    "type": "array",
                                    "items": {
                                          "$ref": "#/components/schemas/StopTimeWithStopDto"
                                    },
                                    "nullable": true
                              }
                        },
                        "additionalProperties": false
                  }
            }
      }
}